Only apply pending promotions if there is a current_user (related to #836)

+++ promo/app/controllers/spree/checkout_controller_decorator.rb
@@ -46,9 +46,11 @@ Spree::CheckoutController.class_eval do
 
   def apply_pending_promotions
     # Apply all promotions that may have been triggered before theorder had been created
-    current_user.pending_promotions.each do |promotional|
-      promotional.promotion.activate(:order => @order, :user => current_user)
-      promotional.destroy
+    if current_user
+      current_user.pending_promotions.each do |promotional|
+        promotional.promotion.activate(:order => @order, :user => current_user)
+        promotional.destroy
+      end
     end
   end
 end
